Luck Rebalance
Technophreak and Luck 9 ended up being absurdly powerful when one considered the additional bonuses for better loot that luck granted (as it ups decon loot success too). This created a scenario where with a lucky roll you could get 11 regular bits of salvage while also rolling 9 times for advanced salvage at a success chance of 52% per 9 rolls. In laymens terms this means you get an average of 10 salvage stacks with 6 regular 4 advanced.

To fix this and address and issue I had with tying it to the technophreak perk directly (you did not get the bonus loot without the perk) luck now functions independently of the perk and at a much more balanced rate.

If you have luck 3 or greater when salvaging a roll will be made of random(0, Luck/4) which means every roll selects a random number between 0 and a maximum of 3 (it rounds up to the nearest hole number) and assigns that as bonus loot modifiers. For basic salvage, this means you can get a maximum of 6 basic salvage through luck alone and 9 salvage if you have technophreak (rolling 9 is unlikely).

For advanced salvage, you get roll attempts at the same rate for regular salvage drops. If you roll 3 with luck 9 and 3 with technophreak you get 7 attempts of advanced salvage on a car wreck. The game then rolls a (25% + Luck * 3) to spawn advanced salvage for the amount. Under the luckiest circumstance, you will attempt 7 rolls at 52% success chance.

This means that luck not only has a direct higher chance to get basic loot, but adds to the roll attempts and the direct odds for salvaged loot.  So it retains power, but is a little less ridiculous.

Next, you have salvaging at a bench. When you salvage tech at a workbench you get 1 drop plus another drop if you have technophreak. A randomized roll is then made with odds of 30% + Luck * 7 (36% min, 93% max) for an additional drop. If this roll succeeds, another roll happens at odds of 5% + Luck * 4 (9% min, 46%) for an additional drop. This means that with luck 9 and technophreak for 1 salvage part you have up to +2 additional loot drops, with an almost guaranteed success at luck 9 for one and an edged coin flip for two.

With luck 2 or 1, you get 0 bonuses to salvage and have an extremely low chance to get bonus loot at workbenches.

tl;dr Technophreak is no longer needed for luck bonuses to salvaging cars/machines. Luck scales more fairly with bonus salvage, luck 8 gives 0-2 bonus and luck 9 gives 0-3. Luck factors more heavily into successful loot when salvaging at a bench. Low luck will have a noticeable difference when salvaging now. You now get a message when you roll bonus loot on the exact amount.